# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Resource Booking API",
        "description": "API для управления бронированием ресурсов.",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Local server"
        }
    ],
    "paths": {
        "/api/bookings": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Создание нового бронирования",
                "operationId": "5391981a6a4e63728ebee2ab5e51cac0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "resource_id",
                                    "user_id",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "resource_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-07-10T10:00:00Z"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-07-10T12:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Успешное бронирование",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/bookings/{id}": {
            "delete": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Удалить бронирование",
                "operationId": "0d3d78ce300894d09e1dc9ef85aa5a22",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID бронирования",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Бронирование удалено"
                    },
                    "404": {
                        "description": "Бронирование не найдено"
                    }
                }
            }
        },
        "/api/resources": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Получить список ресурсов",
                "operationId": "544205ca6e15f97161a1d5b2d6a70ee7",
                "responses": {
                    "200": {
                        "description": "Список ресурсов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ResourceResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Resources"
                ],
                "summary": "Создание нового ресурса",
                "operationId": "62db8e1e6914710d566758560695557f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Conference Room"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "room"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Main hall with projector"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ресурс создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResourceResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/resources/{id}/bookings": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Получить бронирования для ресурса",
                "operationId": "478b32c77d4f2a85c492b1d48da7db1a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID ресурса",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список бронирований",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BookingResource"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "BookingResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "start_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-10T10:00:00Z"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-10T11:00:00Z"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-10T09:00:00Z"
                    },
                    "resource": {
                        "$ref": "#/components/schemas/ResourceResource"
                    }
                },
                "type": "object"
            },
            "ResourceResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Conference Room"
                    },
                    "type": {
                        "type": "string",
                        "example": "room"
                    },
                    "description": {
                        "type": "string",
                        "example": "Main hall with projector"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Bookings",
            "description": "Операции с бронированиями"
        },
        {
            "name": "Resources",
            "description": "Управление ресурсами (комнаты, автомобили и т.д.)"
        }
    ]
}